requirements.txt
asgiref==3.8.1
Django==5.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
gunicorn==23.0.0
djangorestframework==3.16.0
drf-yasg==1.21.10
pytest==8.3.5
pytest-django==4.11.1
